Move GUI class to own file
